% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Activation.R
\name{act.BinReg}
\alias{act.BinReg}
\title{Activation for Binary Regression}
\usage{
act.BinReg(h, model, df = NULL)
}
\arguments{
\item{h}{Linear predictor.}

\item{model}{Selected from among logistic, probit, and robit.}

\item{df}{Degrees of freedom if the robit model is selected.}
}
\description{
Activation for Binary Regression
}
