% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Delta.R
\name{delta.BinReg}
\alias{delta.BinReg}
\title{Delta for Binary Regression}
\usage{
delta.BinReg(h, model, df = NULL)
}
\arguments{
\item{h}{Linear predictor.}

\item{model}{Selected from among logistic, probit, and robit.}

\item{df}{Degrees of freedom if the robit model is selected.}
}
\description{
Delta for Binary Regression
}
