% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Class.R
\docType{class}
\name{fit-class}
\alias{fit-class}
\title{Regression Model}
\description{
Defines the object class returned by fitting functions.
}
\section{Slots}{

\describe{
\item{\code{Model}}{Type of fit.}

\item{\code{Df}}{Degrees of freedom, if using the robit model.}

\item{\code{Coefficients}}{Regression coefficients.}

\item{\code{Eta}}{Linear predictors.}

\item{\code{Information}}{Information.}

\item{\code{Residuals}}{Phenotypic residuals.}
}}

