% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Fit.R
\name{Fit.BinReg}
\alias{Fit.BinReg}
\title{Fit Binary Regression Model}
\usage{
Fit.BinReg(y, X = NULL, model = "logistic", offset = 0, df = NULL,
  sig = 0.05, eps = 1e-08, maxit = 10, report = T)
}
\arguments{
\item{y}{Binary 0/1 outcome vector.}

\item{X}{Numeric model matrix. Include an intercept.}

\item{model}{Selected from among logistic, probit, and robit.}

\item{offset}{Fixed component added to the linear predictor.}

\item{df}{Degrees of freedom, if using the robit model.}

\item{sig}{Significance level, for CIs.}

\item{eps}{Tolerance for Newton-Raphson iterations.}

\item{maxit}{Maximum number of NR iterations.}

\item{report}{Report fitting progress?}
}
\description{
Fit Binary Regression Model
}
\examples{
\dontrun{
set.seed(100);
# Design matrix
X = cbind(1,matrix(rnorm(3e3),nrow=1e3));
# Coefficient
b = c(1,-1,1,0);

# Logistic observations
y = rBinReg(X,b,model="logistic");
# Estimate logistic model
M = Fit.BinReg(y,X,model="logistic");

# Probit observations
y = rBinReg(X,b,model="probit");
# Estimate probit model
M = Fit.BinReg(y,X,model="probit");

# Robit observations
y = rBinReg(X,b,model="robit",df=5);
# Estimate robit model
M = Fit.BinReg(y,X,model="robit",df=5);
}
}
