% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Probit.R
\name{fit.probit}
\alias{fit.probit}
\title{Fit Probit Regression}
\usage{
fit.probit(y, X, sig = 0.05, eps = 1e-08, maxit = 10, report = T)
}
\arguments{
\item{y}{Binary 0/1 outcome vector.}

\item{X}{Numeric model matrix.}

\item{sig}{Significance level, for CIs.}

\item{eps}{Tolerance for Newton-Raphson iterations.}

\item{maxit}{Maximum number of NR iterations.}

\item{report}{Report fitting progress?}
}
\value{
An object of class \code{fit} containing the regression coefficients,
information, and residuals.
}
\description{
Fit Probit Regression
}
\examples{
\dontrun{
set.seed(101);
# Design matrix
n = 1e3;
X = matrix(rnorm(4*n),nrow=n);
# Coefficient
b = c(1,-0.5,-0.5,0);
# Probit outcomes
y = rBinary(X=X,b=b,model="probit");
# Fit probit model
M = fit.probit(y=y,X=X);
show(M);
} 
}
