% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Robit.R
\name{fit.Robit}
\alias{fit.Robit}
\title{Fit Robit Regression}
\usage{
fit.Robit(y, Z, df = 7, alpha = 0.05, eps = 1e-06, maxit = 10)
}
\arguments{
\item{y}{Binary 0/1 outcome vector.}

\item{Z}{Numeric model matrix.}

\item{df}{Degrees of freedom.}

\item{alpha}{Significance level, for CIs.}

\item{eps}{Tolerance for Newton-Raphson iterations.}

\item{maxit}{Maximum number of NR iterations.}
}
\description{
Fit Robit Regression
}
