% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{OLS}
\alias{OLS}
\title{Ordiary Least Squares}
\usage{
OLS(Z, y)
}
\arguments{
\item{Z}{Design matrix.}

\item{y}{Response vector.}
}
\value{
Numeric vector.
}
\description{
Estimates the ordinary least squares coefficient 
\deqn{\hat{\beta}=(Z'Z)^{-1}Z'y}
}
