% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Class.R
\name{predict.fit}
\alias{predict.fit}
\title{Prediction from a Fitted Regression Model}
\usage{
\method{predict}{fit}(object, ..., X = NULL, type = "eta")
}
\arguments{
\item{object}{A \code{fit} object.}

\item{...}{Unused.}

\item{X}{Numeric model matrix. If om}

\item{type}{Either response or eta, the linear predictor.}
}
\value{
A numeric vector of predictions.
}
\description{
Returns the estimated residuals from an object of class \code{fit}.
}
