% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Objective.R
\name{Q.BinReg}
\alias{Q.BinReg}
\title{Objective Function for Binary Regression}
\usage{
Q.BinReg(y, h, model, df = NULL)
}
\arguments{
\item{y}{Outcome vector.}

\item{h}{Linear predictor.}

\item{model}{Selected from among logistic, probit, and robit.}

\item{df}{Degrees of freedom if the robit model is selected.}
}
\description{
Objective Function for Binary Regression
}
