% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataGen.R
\name{rBinReg}
\alias{rBinReg}
\title{Random Binary Outcome}
\usage{
rBinReg(X, b, model = "logistic", df = NULL)
}
\arguments{
\item{X}{Design matrix.}

\item{b}{Regression coefficient.}

\item{model}{Selected from among logistic, probit, and robit.}

\item{df}{Degrees of freedom if the robit model is selected.}
}
\value{
Numeric vector of binary outcomes, one for each row of the design
  matrix.
}
\description{
Simulates outcomes from a binary regression model. Model choices include
logistic, probit, and robit.
}
