% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Class.R
\name{residuals.fit}
\alias{residuals.fit}
\title{Extract Residuals from a Fitted Regression Model}
\usage{
\method{residuals}{fit}(object, ...)
}
\arguments{
\item{object}{A \code{fit} object.}

\item{...}{Unused.}
}
\value{
A numeric vector of residuals.
}
\description{
Returns the estimated residuals from an object of class \code{fit}.
}
