% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Weights.R
\name{weight.BinReg}
\alias{weight.BinReg}
\title{Weights for Binary Regression}
\usage{
weight.BinReg(h, model, df = NULL)
}
\arguments{
\item{h}{Linear predictor.}

\item{model}{Selected from among logistic, probit, and robit.}

\item{df}{Degrees of freedom if the robit model is selected.}
}
\description{
Weights for Binary Regression
}
