% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{WLS}
\alias{WLS}
\title{Weighted Least Squares}
\usage{
WLS(Z, w, y)
}
\arguments{
\item{Z}{Design matrix.}

\item{w}{Weight vector.}

\item{y}{Response vector.}
}
\value{
Numeric vector.
}
\description{
Estimates the weighted least squares coefficient 
\deqn{\hat{\beta}=(Z'WZ)^{-1}Z'Wy}
}
\details{
The weight matrix is assumed diagonal, and is supplied
as a vector.
}
