% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Fit.R
\name{fit.BinReg}
\alias{fit.BinReg}
\title{Fit Binary Regression Model}
\usage{
fit.BinReg(y, X = NULL, model = "logistic", df = NULL, sig = 0.05,
  eps = 1e-08, report = T)
}
\arguments{
\item{y}{Binary 0/1 outcome vector.}

\item{X}{Numeric model matrix. Include an intercept.}

\item{model}{Selected from among logistic, probit, and robit.}

\item{df}{Degrees of freedom.}

\item{sig}{Significance level, for CIs.}

\item{eps}{Tolerance for Newton-Raphson iterations.}

\item{report}{Report fitting progress?}

\item{maxit}{Maximum number of NR iterations.}
}
\description{
Fit Binary Regression Model
}
