% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Probit.R
\name{fit.Probit}
\alias{fit.Probit}
\title{Fit Probit Regression}
\usage{
fit.Probit(y, Z, alpha = 0.05, eps = 1e-06, maxit = 10)
}
\arguments{
\item{y}{Binary 0/1 outcome vector.}

\item{Z}{Numeric model matrix.}

\item{alpha}{Significance level, for CIs.}

\item{eps}{Tolerance for Newton-Raphson iterations.}

\item{maxit}{Maximum number of NR iterations.}
}
\description{
Fit Probit Regression
}
